Fix some tests

+++ totaleconomy-sponge/src/main/java/com/ericgrandt/totaleconomy/impl/UniqueAccountImpl.java
@@ -234,7 +234,7 @@ public class UniqueAccountImpl implements UniqueAccount {
                 accountId,
                 UUID.fromString(to.identifier()),
                 currencyDto.id(),
-                amount.doubleValue()
+                amount
             );
         } catch (SQLException e) {
             logger.error(
